+++ mail/proton-mail-uniffi/CHANGELOG.md
@@ -5,8 +5,72 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased] - 2024-00-00
+
+## [0.10.34] - 2024-07-22
+
+### Changed
+
+- `DecryptedMessage` now reads mail settings from the database.
+
+## [0.10.33] - 2024-07-19
+
+### Added
+
+- The following functions have been exposed:
+    - avatar_information_from_name_and_email
+    - avatar_information_from_message_addresses
+    - avatar_information_from_message_address
+- Disable/Enable remote image in HTML content in `DecryptedMessageBody`.
+- Add `rust_log_*` functions to log into the rust log.
+
+### Changed
+
+- Renamed `DecryptedMessageBody` into `DecryptedMessage`.
+
+## [0.10.32] - 2024-07-15
+
+### Added
+
+- Strip UTM parameters from HTML content
+- [iOS] Inject viewport metadata for web view
+
+## [0.10.31] - 2024-07-02
+
+### Fixed
+
+- Ensure mail cache path is unique per user.
+- Fix conversation message selection.
+- Use Get with query parameters for message metadata.
+
+## [0.10.30] - 2024-06-21
+
+## Changed
+
+- Update `MailUserSession::filter_converstions` to require a label id for context.
+
+## [0.10.29] - 2024-06-21
+
+### Changed
+
+- Adds additional debug logs.
+
+## [0.10.28] - 2024-06-18
+
+### Fixed
+
+- Correctly initialize address id in attachments when created from message data.
+
+## [0.10.27] - 2024-06-12
+
+### Changed
+
+- Message conversation id is no longer optional.
+- The following functions now download their respective content if not available
+    - `MailUserSession::conversation_with_remote_id`
+    - `MailUserSession::conversation_with_id_and_context`
+    - `MailUserSession::conversation_with_id_with_all_mail_context`
+    - `MailUserSession::message_metadata_with_remote_id`
 
 ## [0.10.26] - 2024-06-10